backend/pyproject.toml
[project]
name = "vocabulary-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
    "python-dotenv>=1.0.0",
    "PyJWT>=2.8.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "httpx>=0.27.0",
    "ruff>=0.4.0",
    "pre-commit>=3.7",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.coverage.run]
source = ["."]
omit = ["tests/*", ".venv/*"]
